fix: use workflow object for rendering sample tags
