+++ src/js/samples/types.ts
@@ -1,0 +1,12 @@
+export enum WorkflowState {
+    COMPLETE = "complete",
+    PENDING = "pending",
+    NONE = "none",
+    INCOMPATIBLE = "incompatible",
+}
+
+export type SampleWorkflows = {
+    aodp: WorkflowState;
+    nuvs: WorkflowState;
+    pathoscope: WorkflowState;
+};
